refactor: centralize ssl configuration
