+++ src/ume/utils.py
@@ -1,0 +1,17 @@
+from typing import Dict
+import os
+
+
+def ssl_config() -> Dict[str, str]:
+    """Return Kafka SSL configuration if cert env vars are set."""
+    ca = os.environ.get("KAFKA_CA_CERT")
+    cert = os.environ.get("KAFKA_CLIENT_CERT")
+    key = os.environ.get("KAFKA_CLIENT_KEY")
+    if ca and cert and key:
+        return {
+            "security.protocol": "SSL",
+            "ssl.ca.location": ca,
+            "ssl.certificate.location": cert,
+            "ssl.key.location": key,
+        }
+    return {}
