+++ tests/test_demo_scripts.py
@@ -1,19 +1,20 @@
 from types import SimpleNamespace
 
 
-from ume import consumer_demo, producer_demo
+from ume import producer_demo
+from ume.utils import ssl_config
 
 
 def test_consumer_ssl_config(monkeypatch):
     monkeypatch.delenv("KAFKA_CA_CERT", raising=False)
     monkeypatch.delenv("KAFKA_CLIENT_CERT", raising=False)
     monkeypatch.delenv("KAFKA_CLIENT_KEY", raising=False)
-    assert consumer_demo.ssl_config() == {}
+    assert ssl_config() == {}
 
     monkeypatch.setenv("KAFKA_CA_CERT", "ca")
     monkeypatch.setenv("KAFKA_CLIENT_CERT", "cert")
     monkeypatch.setenv("KAFKA_CLIENT_KEY", "key")
-    cfg = consumer_demo.ssl_config()
+    cfg = ssl_config()
     assert cfg["security.protocol"] == "SSL"
     assert cfg["ssl.ca.location"] == "ca"
     assert cfg["ssl.certificate.location"] == "cert"
@@ -24,12 +25,12 @@ def test_producer_ssl_config(monkeypatch):
     monkeypatch.delenv("KAFKA_CA_CERT", raising=False)
     monkeypatch.delenv("KAFKA_CLIENT_CERT", raising=False)
     monkeypatch.delenv("KAFKA_CLIENT_KEY", raising=False)
-    assert producer_demo.ssl_config() == {}
+    assert ssl_config() == {}
 
     monkeypatch.setenv("KAFKA_CA_CERT", "ca")
     monkeypatch.setenv("KAFKA_CLIENT_CERT", "cert")
     monkeypatch.setenv("KAFKA_CLIENT_KEY", "key")
-    cfg = producer_demo.ssl_config()
+    cfg = ssl_config()
     assert cfg["security.protocol"] == "SSL"
     assert cfg["ssl.ca.location"] == "ca"
     assert cfg["ssl.certificate.location"] == "cert"
